Switch to long in timeouts

+++ halflife-core/src/java/halflife/bus/channel/ConsumingChannel.java
@@ -7,7 +7,7 @@ import java.util.concurrent.TimeUnit;
 public interface ConsumingChannel<T> {
 
   public T get();
-  public T get(int time, TimeUnit timeUnit) throws InterruptedException;
+  public T get(long time, TimeUnit timeUnit) throws InterruptedException;
   public AnonymousStream<T> stream();
 
 }
